Remove attempt to build path when following links
Commit 3464826 added attempt to construct a path to the referent of a
symlink, which was desired to make following symlinks work on Windows.
However, paths can be constructed (easily done w/ multiple symlinks)
that cause this approach to fail, at least for Unix.

Since Stat() will traverse the paths with an ending symlink (including
any intermediate symlinks in the path), manipulating the path is
unneeded on Unix.  For now, remove Readlink() and path construction and
just rely on Stat(), going back to the original behavior.

The TestWalkSymbolicRelativeLinkChain has an example directory tree that
fails on Unix before this commit, which was constructed with the
following shell commands:

mkdir a b
touch z
ln -s ../b a/x
ln -s ../z b/y

Note that Go has an EvalSymlinks() func which is designed to return a
direct pathname without the symlinks, and which could be useful for
fixing the problems that originally caused this (now removed) code to be
added.  It has a cost, in that it does an Lstat() on every component of
a path, to find the symlinks.  Separating the helper func into os
specific files could at least restrict it's use to Windows only.