Use EvalSymlinks() when FollowSymbolicLinks is set
Commit 3464826 added some symlink content reading (via Readlink()) and
pathname manipulation to purportedly allow symlinks to work on Windows
(but which caused problems with certain symlink containing paths on
Unix).

The Go EvalSymlinks() func appears to have been designed to help with
this kind of situation, so call it on paths that we know end with
symlinks, and may point to a directory.