test ErrorCallback responses without noaccess
ErrorCallback is now triggered by attempting to visit the symbolic
link that points to a missing file system entry, rather than a file
for which the testing user does not have permissions.  This is because
on some platforms it was difficult to setup and teardown test
scaffolding with a directory that had restricted permissions.