Scanner, ReadDirents, ReadDirnames scratch buffer
While updating Walk to work with Scanner, race condition identified
with Scanner's use of scratch buffer. This confirms my previous
suspicions, and I decided to go ahead with this change which I had
already considered before but dismissed, namely always creating a
scratch buffer for each Scanner instance, albeit a smaller scratch
buffer.

This is because Scanner only needs to really read a single item at a
time from the file system. While we might be able to select a smaller
scratch buffer size than the OS pagesize, I want to make sure this
does not fail when a really long file name is in a directory.