Options.AllowNonDirectory
AllowNonDirectory causes Walk to bypass the check that ensures it is
being called on a directory node, or when FollowSymbolicLinks is true, a
symbolic link that points to a directory. Leave this value false to have
Walk return an error when called on a non-directory. Set this true to
have Walk run even when called on a non-directory node.