find-fast provides `--skip PATTERN` to skip entries (typo corrected)
Ref: #53

Returning the special value `filepath.SkipDir` as an error token is
supported by `filepath.Walk` and by `godirwalk.Walk` to configure the
walk function to not descend further in the directory hierarchy of the
current file system entry.

This change updates the example find-fast program to demonstrate use
of this token error value.