fix: ignore EOF error on Windows directory scanning
In #68, the scanner now returns any error that was encountered. On
Windows, the call to Readdir will return an EOF when it is done reading
a directory.

I believe the desired behavior is to throw an error on any actual
errors, but in the case of EOF, continue on to the next directory. As
such, check for the EOF and return no error while releasing resources.

fixes: #70